server-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF>=1.24.2
python-docx==1.1.0
Pillow==10.1.0
aiofiles==23.2.1

# Optional for OCR (uncomment if needed)
# pytesseract==0.3.10